crud de usuário, migration e tabela relacional admin_evento

+++ database/migrations/2021_09_18_001103_create_doadors_table.php
@@ -19,7 +19,7 @@ class CreateDoadorsTable extends Migration
             $table->date('dt_nascimento');
             $table->string('endereco', 100);
             $table->integer('num_residencia');
-            $table->integer('telefone');
+            $table->bigInteger('telefone');
             $table->string('cidade_doacao', 50);
             $table->string('tipo_sanguineo');
            // $table->bigInteger('usuario_id')->unsigned();
